All tracks to be generated with or without client_secret
